# app/frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/aid/campaigns/{id}": {
      "patch": {
        "description": "Modifies an existing campaign. Only provided fields will be updated.",
        "operationId": "AidController_updateCampaign_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign updated successfully."
          },
          "400": {
            "description": "Invalid update data."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a campaign",
        "tags": [
          "Aid"
        ]
      },
      "delete": {
        "description": "Soft-archives a campaign, making it invisible to standard listings.",
        "operationId": "AidController_archiveCampaign_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign archived successfully."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Archive a campaign",
        "tags": [
          "Aid"
        ]
      }
    },
...
    "/api/v1/campaigns": {
      "post": {
        "operationId": "CampaignsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created successfully."
          },
          "400": {
            "description": "Invalid input parameters or malformed request body."
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          },
          "403": {
            "description": "Access denied - insufficient permissions for this operation."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a campaign",
        "tags": [
          "Campaigns"
        ]
      },
      "get": {
        "description": "Retrieves campaigns. NGO operators only see their own organization's campaigns.",
        "operationId": "CampaignsController_list_v1",
        "parameters": [
          {
            "name": "includeArchived",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return (default: 25, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor for pagination",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResult"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CreateCampaignDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List all campaigns",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}": {
      "get": {
        "description": "Retrieves metadata and status for a specific campaign.",
        "operationId": "CampaignsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign details retrieved successfully."
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get campaign details",
        "tags": [
          "Campaigns"
        ]
      },
      "patch": {
        "description": "Modifies existing campaign properties. Only provided fields are updated.",
        "operationId": "CampaignsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign updated successfully."
          },
          "400": {
            "description": "Invalid update data or malformed request body."
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          },
          "403": {
            "description": "Access denied - insufficient permissions."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a campaign",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/archive": {
      "patch": {
        "description": "Marks a campaign as archived. Archived campaigns are hidden from general listings.",
        "operationId": "CampaignsController_archive_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign archived successfully."
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          },
          "403": {
            "description": "Access denied - insufficient permissions."
          },
          "404": {
            "description": "The specified campaign was not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Archive campaign (soft archive)",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/balance": {
      "get": {
        "description": "Returns the current locked, disbursed, and available budget for a campaign. Locked balance accounts for all active (non-cancelled, non-disbursed) claims. Cancelled claims release their locked amount back to available.",
        "operationId": "CampaignsController_getBalance_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign balance retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "campaignId": {
                      "type": "string"
                    },
                    "budget": {
                      "type": "number",
                      "description": "Total campaign budget."
                    },
                    "lockedAmount": {
                      "type": "number",
                      "description": "Amount locked by active claims."
                    },
                    "disbursedAmount": {
                      "type": "number",
                      "description": "Amount already disbursed."
                    },
                    "availableBudget": {
                      "type": "number",
                      "description": "Remaining available budget."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - operator role required."
          },
          "404": {
            "description": "Campaign not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get campaign balance summary",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/budget-summary": {
      "get": {
        "description": "Returns the total, locked, disbursed, and available budget for a campaign.",
        "operationId": "CampaignsController_getBudgetSummary_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign budget summary retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "campaignId": {
                      "type": "string"
                    },
                    "budget": {
                      "type": "number"
                    },
                    "locked": {
                      "type": "number"
                    },
                    "disbursed": {
                      "type": "number"
                    },
                    "available": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get campaign budget summary",
        "tags": [
          "Campaigns"
        ]
      }
    },
    "/api/v1/campaigns/export": {
      "get": {
        "description": "Exports campaign summaries as CSV with support for date range, status, organization, and pagination filters. Excludes sensitive internal metadata and deleted records.",
        "operationId": "CampaignsController_exportCampaigns_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 50, max: 200)",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {}
          },
          {
            "name": "ngoId",
            "required": false,
            "in": "query",
            "description": "NGO ID filter",
            "schema": {}
          },
          {
            "name": "orgId",
            "required": false,
            "in": "query",
            "description": "Organization ID filter",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Campaign status filter",
            "schema": {}
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "End date (ISO string)",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Start date (ISO string)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns exported successfully.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials."
          },
          "403": {
            "description": "Access denied - operator or admin role required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export campaigns as CSV",
        "tags": [
          "Campaigns"
        ]
      }
    },
...
    "/api/v1/entity-linking/campaign/{campaignId}": {
      "get": {
        "description": "Retrieve all entity links associated with a specific campaign",
        "operationId": "EntityLinkingController_getLinksByCampaign_v1",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "organization",
                "location",
                "asset",
                "project"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get entity links by campaign",
        "tags": [
          "Entity Linking"
        ]
      }
    },
...
  },
  "info": {
    "title": "ChainForge API",
    "description": "API documentation for ChainForge platform - Humanitarian aid and verification system\n\n## API Versioning\n\nThis API uses URI-based versioning. The current version is **v1**.\n\n### Version Format\nAll endpoints are prefixed with the version number: `/api/v1/...`\n\n### Supported Versions\n| Version | Status | Description |\n|---------|--------|-------------|\n| v1 | Current | Active version with full support |\n\n### Deprecation Policy\n- Deprecated endpoints will be marked with `@Deprecated` in the documentation\n- Deprecated versions will be supported for at least 6 months after deprecation notice\n- Clients will receive deprecation warnings via the `Sunset` HTTP header\n- Migration guides will be provided for major version changes\n\n### Future Versions\nWhen new versions are released:\n- New endpoints will be available at `/api/v2/...`, etc.\n- Previous versions remain accessible during the deprecation period\n- Clients should monitor the API documentation for version updates",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Staging"
    },
    {
      "url": "https://api.chainforge.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Enter JWT token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external access"
      }
    },
    "schemas": {
...
      "CampaignStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "paused",
          "completed",
          "archived"
        ],
        "description": "Current status of the campaign."
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Campaign title/name.",
            "example": "Winter Relief 2026"
          },
          "budget": {
            "type": "number",
            "description": "Total budget allocated to the campaign.",
            "example": 25000.5,
            "minimum": 0
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary campaign metadata (e.g., region, location, target audience).",
            "example": {
              "region": "Lagos",
              "partner": "NGO-A",
              "notes": "Phase 1"
            },
            "additionalProperties": true
          },
          "status": {
            "description": "Current status of the campaign.",
            "example": "draft",
            "allOf": [
              {
                "$ref": "#/components/schemas/CampaignStatus"
              }
            ]
          }
        },
        "required": [
          "name",
          "budget"
        ]
      },
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The array of records for the current page.",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Cursor to retrieve the next page of results, or null if there is no next page."
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated campaign title/name.",
            "example": "Winter Relief 2026 - Extended"
          },
          "budget": {
            "type": "number",
            "description": "Updated campaign budget.",
            "example": 30000,
            "minimum": 0
          },
          "metadata": {
            "type": "object",
            "description": "Updated campaign metadata.",
            "example": {
              "region": "Lagos",
              "partner": "NGO-B"
            },
            "additionalProperties": true
          },
          "status": {
            "description": "Updated campaign status.",
            "example": "active",
            "allOf": [
              {
                "$ref": "#/components/schemas/CampaignStatus"
              }
            ]
          }
        }
      },
...
    }
  }
}